% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mids_update.R
\name{as.mids.update}
\alias{as.mids.update}
\title{Slightly edited version of as.mids() from mice package}
\usage{
as.mids.update(data, .imp = 1, .id = 2)
}
\arguments{
\item{data}{A multiply imputed data set in long format}

\item{.imp}{Mandatory column indicator for the multiple imputation stream, where 0 indicates the
incomplete data and 1 through m indicate the m multiple imputation streams. Default is 1.}

\item{.id}{Optional column indicator for the row numbers. Default is 2.}
}
\description{
The original function relies on data being sorted by .imp to work properly. This is problematic
in data sets created manually - say, calculating summary statistics using original + imputed
data, where the final data set we want to combine does not have an .imp column. Original function
by Stef van Buuren, with detective work and changes by Cole Beck. See
\href{https://github.com/jenniferthompson/TBIMort/blob/master/tbi_mortality_riskfactors.Rmd}{this
 analysis} for a full example.
}

