% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nb_ratioci.R
\name{calc_nb_ratioci}
\alias{calc_nb_ratioci}
\title{Calculate incidence rate ratio and 95% confidence limits from a negative binomial regression
model for a given comparison in a continuous predictor variable.}
\usage{
calc_nb_ratioci(nbObj, nbCoefs, nbVcov, predVar, adjustTo = NULL, df,
  alpha = 0.05)
}
\arguments{
\item{nbObj}{Model fit of class glm.nb, or mice::mira object fit using glm.nb.}

\item{nbCoefs}{Vector of coefficients for a glm.nb model.}

\item{nbVcov}{Variance-covariance matrix for a glm.nb model.}

\item{predVar}{Character string; name of main predictor variable.}

\item{adjustTo}{Numeric vector of length 2; values to adjust predVar to. Defaults to c(25th,
75th percentiles).}

\item{df}{Data frame used to get defaults for adjustTo and to calculate nonlinear terms if
applicable.}

\item{alpha}{Alpha level for confidence limits. Defaults to 0.05.}
}
\value{
data.frame with one row, containing columns for point estimate, confidence limits, and
reference and comparison values used.
}
\description{
Currently doesn't handle interaction terms or categorical predVars.
}
\examples{

## Create data frame
df <- data.frame(y = round(rexp(n = 100, rate = 0.5)),
                 v1 = sample(1:5, size = 100, replace = TRUE),
                 v2 = rnorm(n = 100))

## Fit negative binomial model
mymod <- MASS::glm.nb(y ~ v1 * v2, data = df)

calc_nb_ratioci(mymod, predVar = 'v1', df = df)

}
\seealso{
\code{\link[MASS]{glm.nb}}; \code{\link[mice]{is.mice}}, \code{\link[mice]{pool}};
\code{\link[Hmisc]{rcspline.eval}} for nonlinear terms.
}

