% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_countprocess_data.R
\name{create_countprocess_data}
\alias{create_countprocess_data}
\title{Create count process data set, eg, for use in time-dependent Cox proportional hazards models.}
\usage{
create_countprocess_data(org.data, id.var, record.var, time.var, event.var,
  event.string, data.set, out.strings = c("Alive through end of interval",
  "Died at end of interval"))
}
\arguments{
\item{org.data}{Original data frame with multiple records per patient}

\item{id.var}{Variable name for subject identifier (eg, ID)}

\item{record.var}{Variable name for record identifier (eg, study day)}

\item{time.var}{Variable indicating end of records (eg, day of death)}

\item{event.var}{Variable indicating whether event happened (eg, "died.yn")}

\item{event.string}{Level of event variable indicating whether event happened (eg, "yes")}

\item{data.set}{Data set (one record/pt) containing time/event variables}

\item{out.strings}{Labels for 0/1 values of outcome variable}
}
\value{
data.frame containing reference, comparison, effect, lower and upper confidence limits,
variable name and indicator for whether row contains reference:reference comparison.
}
\description{
Note that Zhiguo (Alex) Zhao wrote the majority of this code, with later edits by Cole Beck and
minor ones by me. It is here for my convenience, since Alex shared it with me and to my knowledge
it is not available elsewhere, not in an attempt to take credit.
}
\examples{

## Create longitudinal data frame
longdf <- data.frame(id = c(rep(1, 8), rep(2, 10)),
                     day = c(1:8, 1:10),
                     v1 = sample(1:5, size = 18, replace = TRUE),
                     v2 = sample(LETTERS[1:5], size = 18, replace = TRUE))

## Create data frame with event status and times
eventdf <- data.frame(id = 1:2,
                      died = c('Yes', 'No'),
                      days.deathorend = c(8, 10))

## Create count process data set
create_countprocess_data(org.data = longdf,
                         id.var = 'id',
                         record.var = 'day',
                         time.var = 'days.deathorend',
                         event.var = 'died',
                         event.string = 'Yes',
                         data.set = eventdf)
}
\seealso{
\href{https://cran.r-project.org/web/packages/survival/vignettes/timedep.pdf}{Terry
Therneau et al's vignette on time-dependent Cox models}, \href{https://cran.r-project.org/web/packages/survival/}{survival package}.
}

