% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatp.R
\name{formatp}
\alias{formatp}
\title{Format a p-value, somewhat per NEJM style.}
\usage{
formatp(p)
}
\arguments{
\item{p}{Numeric; p-value to be rounded and formatted}
}
\value{
Character string; <0.0001, <0.001, or rndformat(p), depending on the value of p.
}
\description{
Format a p-value, somewhat per NEJM style.
}
\examples{
formatp(0.000214)
formatp(0.28723)

}

