% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnostics.R
\name{lm_diagnostics}
\alias{lm_diagnostics}
\title{Check assumptions of heteroscedasticity and normally distributed errors for a linear regression
model fit.}
\usage{
lm_diagnostics(lmObj, outcomeString = "Outcome", titleString)
}
\arguments{
\item{lmObj}{Model fit of class `lm`.}

\item{outcomeString}{String to identify outcome in X axis label. Defaults to "Outcome."}

\item{titleString}{String to add to plot titles.}
}
\description{
Check assumptions of heteroscedasticity and normally distributed errors for a linear regression
model fit.
}
\examples{

## Linear regression model
mymod <- lm(Sepal.Width ~ Sepal.Length, data = iris)

lm_diagnostics(mymod, outcomeString = 'Sepal.Width', titleString = 'Sepal.Width vs Sepal.Length')

}

