% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest_removeTerm.R
\name{lrtest_removeTerm}
\alias{lrtest_removeTerm}
\alias{lrtest_removeTerm.default}
\alias{lrtest_removeTerm.mira}
\title{Likelihood ratio p-values for full model vs model with one term (or consecutive terms) removed}
\usage{
lrtest_removeTerm(orgModel, removeTerm, ...)

\method{lrtest_removeTerm}{default}(orgModel, removeTerm)

\method{lrtest_removeTerm}{mira}(orgModel, removeTerm, miceObjName)
}
\arguments{
\item{orgModel}{Model object containing full model.}

\item{removeTerm}{Character string of term(s) to be removed.}

\item{...}{Additional arguments specific to method.}

\item{miceObjName}{Character string; name of `mice` object.}

\item{orgModel}{Model object containing full model.}

\item{removeTerm}{Character string of term(s) to be removed.}

\item{orgModel}{Model object containing full model.}

\item{removeTerm}{Character string of term(s) to be removed.}
}
\value{
List with following values from lrtest() or pool.compare(): test degrees of freedom,
X^2^ value, and p-value.
}
\description{
Runs model without `removeTerm` and performs likelihood ratio test for full model vs nested model.
`function(variable)` - eg, `rcs(age, 3)` - is considered one term, as are all levels of a
categorical variable.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method.

\item \code{mira}: Method for models fit with the mice package.
}}
\examples{

## Fit full model using iris data
fullModel <- lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width, data = iris)

## What about just the effect of Petal.Length?
lrtest_removeTerm(fullModel, 'Petal.Length')

## What about the effect of both Petal variables?
lrtest_removeTerm(fullModel, 'Petal.Length + Petal.Width')


## Add missingness to iris data
iris_missing <- iris
iris_missing[sample(1:nrow(iris), size = round(nrow(iris_missing) / 10)), 'Sepal.Width'] <- NA

## Create mice object
iris_mice <- mice(iris_missing)

#' ## Fit full model using iris data
fullModel <- with(iris_mice, lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width))

## What about just the effect of Petal.Length?
lrtest_removeTerm(fullModel, 'Petal.Length', 'iris_mice')

## What about the effect of both Petal variables?
lrtest_removeTerm(fullModel, 'Petal.Length + Petal.Width', 'iris_mice')
}
\seealso{
\code{lrtest_removeTerm.default, lrtest_removeTerm.mira}.

\code{\link[lmtest]{lrtest}}.

\code{\link[mice]{pool.compare}}.
}

