% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lrpval.R
\name{model_lrpval.mira}
\alias{model_lrpval.mira}
\title{Likelihood ratio p-values for full model vs nested model, both fit with mice objects}
\usage{
\method{model_lrpval}{mira}(orgModel, removeTerm, miceObjName)
}
\arguments{
\item{orgModel}{Model object containing full model.}

\item{removeTerm}{Character string of term(s) to be removed.}

\item{miceObjName}{Character string; name of `mice` object.}
}
\value{
Numeric value; p-value from `mice::pool.compare()`.
}
\description{
Runs model without `removeTerm` using a `mice` object and performs likelihood ratio test for full
model vs nested model. `function(variable)` - eg, `rcs(age, 3)` - is considered one term, as are
all levels of a categorical variable.
}
\examples{

## Add missingness to iris data
iris_missing <- iris
iris_missing[sample(1:nrow(iris), size = round(nrow(iris_missing) / 10)), 'Sepal.Width'] <- NA

## Create mice object
iris_mice <- mice(iris_missing)

#' ## Fit full model using iris data
fullModel <- with(iris_mice, lm(Sepal.Length ~ Sepal.Width + Petal.Length + Petal.Width))

## What about just the effect of Petal.Length?
model_lrpval(fullModel, 'Petal.Length', 'iris_mice')

## What about the effect of both Petal variables?
model_lrpval(fullModel, 'Petal.Length + Petal.Width', 'iris_mice')
}
\seealso{
\code{\link[mice]{pool.compare}}.
}

