% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lrpval.R
\name{model_lrpval}
\alias{model_lrpval}
\title{Likelihood ratio p-values for full model vs model with one term removed
Runs model without `removeTerm` and performs likelihood ratio test for full model vs nested model.
`function(variable)` - eg, `rcs(age, 3)` - is considered one term, as are all levels of a
categorical variable.}
\usage{
model_lrpval(orgModel, removeTerm, ...)
}
\arguments{
\item{orgModel}{Model object containing full model.}

\item{removeTerm}{Character string of term(s) to be removed.}

\item{...}{Additional arguments specific to method.}
}
\value{
Numeric value; p-value from `lrtest()` or `pool.compare()`.
}
\description{
Likelihood ratio p-values for full model vs model with one term removed
Runs model without `removeTerm` and performs likelihood ratio test for full model vs nested model.
`function(variable)` - eg, `rcs(age, 3)` - is considered one term, as are all levels of a
categorical variable.
}
\seealso{
\code{model_lrpval.default, model_lrpval.mira}.
}

