% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Put multiple ggplot plots on a single page.}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
This function was written completely by Winston Chang and lifted from the Cookbook for R (see
link below).
}
\seealso{
\href{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}{Original
code from Winston Chang's Cookbook for R}.


####################################################################################################
}

