% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_calc_comparisons.R
\name{rms_calc_comparisons}
\alias{rms_calc_comparisons}
\title{Calculate estimates (XB or hazard/odds ratios, depending on model type) for specified values of
a numeric model covariate, or all levels of a factor covariate, vs a reference level. Store in a
data frame.}
\usage{
rms_calc_comparisons(rmsObj, getRatios, df, vname, refVal, compVals,
  nUnique = 10, compQuant = c(0.1, 0.25, 0.5, 0.75, 0.9), rndRC = 2)
}
\arguments{
\item{rmsObj}{Model fit object of class rms.}

\item{getRatios}{Indicator for whether to calculate ratios (exp(XB)) vs estimates on XB scale.
Defaults to TRUE if rmsObj is from cph() or lrm().}

\item{df}{Data frame from which to get variable class information and, if needed, reference
and comparison values.}

\item{vname}{String; name of variable in both names(df) and a covariate in rmsObj.}

\item{refVal}{Value or factor level to set reference level of df[,vname] to. Default is
median (numeric) or first level (factor).}

\item{compVals}{Numeric; value(s) to set comparison levels of vname to, if df[,vname] is numeric.}

\item{nUnique}{Integer; if compVals are not specified and number of unique values of variable
is < nUnique, will calculate estimates for every unique value vs reference.}

\item{compQuant}{Numeric vector; if compVals are not specified, gets these quantiles from df to
use as default comparison values. Defaults to c(0.1, 0.25, 0.5, 0.75, 0.9).}

\item{rndRC}{Integer; number of digits to round reference, comparison columns to for numeric
variables. Defaults to 2.}
}
\value{
data.frame containing reference, comparison, effect, lower and upper confidence limits,
variable name and indicator for whether row contains reference:reference comparison.
}
\description{
Calculate estimates (XB or hazard/odds ratios, depending on model type) for specified values of
a numeric model covariate, or all levels of a factor covariate, vs a reference level. Store in a
data frame.
}
\examples{
## Fit linear regression using ols()
mymod <- ols(Sepal.Length ~ Species + Sepal.Width, data = iris)

## Set datadist
dd.iris <- datadist(iris)
options(datadist = 'dd.iris')

## Continuous covariate, comparing all quantiles to median by default
rms_calc_comparisons(mymod, vname = 'Sepal.Width', df = iris)

## Categorical covariate, comparing all levels to reference
rms_calc_comparisons(mymod, vname = 'Species', df = iris)

}
\seealso{
\code{\link[rms]{ols}}, \code{\link[rms]{lrm}}, \code{\link[rms]{cph}},
\code{\link[rms]{Gls}}, \code{\link[rms]{summary.rms}}.
}

