% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_estimate_with_int.R
\name{rms_estimate_with_int}
\alias{rms_estimate_with_int}
\title{Function to get estimates, on original or ratio scale, for a main variable at a given level of
another variable using summary.rms(). Returns a tidy data frame with columns for main and
interacting variable names and levels and numeric columns for effects and CIs.}
\usage{
rms_estimate_with_int(rmsObjName, estVar, estVals, intVar, intAdjust, getRatios)
}
\arguments{
\item{rmsObjName}{Character string; name of model fit object of class rms}

\item{estVar}{Character string; name of variable for which we want estimate/CI}

\item{estVals}{Vector of length 2; comparison wanted for main variable. Default = datadist() defaults.}

\item{intVar}{Character string; name of interacting variable}

\item{intAdjust}{Character or numeric value; value to adjust intVar to. Default = datadist() default.}

\item{getRatios}{Indicator for whether to calculate ratios (exp(XB)) vs estimates on XB scale.
Defaults to TRUE if get(rmsObjName) is from cph() or lrm().}
}
\value{
data.frame containing reference, comparison, effect, lower and upper confidence limits,
variable name and indicator for whether row contains reference:reference comparison.
}
\description{
This will often be used in conjunction with \code{lapply} or \code{purrr::pmap}. For an example,
see chunk \code{death_inhosp_plot} in \href{https://github.com/jenniferthompson/MotoricSubtypes/blob/master/brain_motoric_subtypes.Rmd}{this Rnw file}
(function here is called \code{estimate.with.interaction.rms}).
}
\examples{
## Fit logistic regression using lrm()
mymod <- lrm(Species ~ Sepal.Length * Sepal.Width, data = iris)

## Set datadist
dd.iris <- datadist(iris)
options(datadist = 'dd.iris')

## Continuous covariate, comparing all quantiles to median by default
rms_estimate_with_int('mymod', estVar = 'Sepal.Length', intVar = 'Sepal.Width')

}
\seealso{
\code{\link[rms]{ols}}, \code{\link[rms]{lrm}}, \code{\link[rms]{cph}},
\code{\link[rms]{Gls}}, \code{\link[rms]{summary.rms}}, \code{\link[purrr]{map}},
\code{\link[purrr]{pmap}}.
}

