% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_model_results.R
\name{rms_model_results}
\alias{rms_model_results}
\title{Combine information from summary.rms(), anova.rms(), and other rms object info to create a
single tidy data.frame of model results that could be printed in a report.}
\usage{
rms_model_results(rmsObj, labeldf = NULL, shortLabs = NULL, rndDigits = 2,
  rndTestStat = 1, rndInfo = rndDigits, printFormat = c("plain",
  "markdown", "latex"), rmRows = c("nl.int", "nl", "int", "none"), addInfo)
}
\arguments{
\item{rmsObj}{Object of class rms}

\item{labeldf}{Data frame from which to get variable labels. Defaults to NULL, in which case
variable names will be used.}

\item{shortLabs}{Named vector of variable labels to replace in interaction rows. Must be in
format c("variable name" = "shortened label").}

\item{rndDigits}{Number of digits to round reference, comparison, result and CI values to.
Defaults to 2.}

\item{rndTestStat}{Number of digits to round test statistic to. Defaults to 1.}

\item{printFormat}{How to format results for printing. Defaults to plain text; LaTeX or Markdown
available.}

\item{rmRows}{Which sets of rows to exclude from final model results. For example, it is more
appropriate to look at the test for total nonlinearity to determine whether to keep these terms
than to look at individual variables' tests for nonlinearity. Example 2: in cases like nonlinear
interactions, some rows of output can be redundant. Options: nl (removes rows labeled
Nonlinear [Interaction] (...) in anova.rms()); int (removes rows labeled All Interactions in
anova.rms()); nl.int (removes all of the above); none (keeps all rows). Defaults to nl.int.}

\item{rndModInfo}{Number of digits to round model LR, R2, etc to. Defaults to rndDigits.}

\item{whichInfo}{Vector of elements of rmsObj$stats to include. Defaults to model likelihood
ratio, R^2, C, Dxy, as available. (Model df are included in the main table.)}
}
\value{
data.frame containing quantity descriptions, reference and comparison values,
coefficient or ratio and 95% CI, test statistic, df, and p-value, along with observations and
model statistics.
}
\description{
Combine information from summary.rms(), anova.rms(), and other rms object info to create a
single tidy data.frame of model results that could be printed in a report.
}
\examples{

## Fit linear regression model using ols()
mymod <- ols(mpg ~ cyl * wt, data = mtcars)

## Make sure to set datadist; summary.rms() requires this
dd.mtcars <- datadist(mtcars)
options(datadist = 'dd.mtcars')

## Defaults, using variable names
rms_model_results(mymod)

## Defaults, using variable labels
label(mtcars$cyl) <- 'Number of cylinders'
label(mtcars$wt) <- 'Weight (1000 lbs)'
rms_model_results(mymod, labeldf = mtcars)

## Use shortened variable labels for interaction row
rms_model_results(mymod, labeldf = mtcars, shortLabs = c('cyl' = 'Cylinders', 'wt' = 'Weight'))

## Use shortened variable labels, format for Markdown
rms_model_results(mymod, labeldf = mtcars, shortLabs = c('cyl' = 'Cylinders', 'wt' = 'Weight'), printFormat = 'markdown')

}
\seealso{
\code{\link[rms]ols()}, \code{\link[rms]lrm()}, \code{\link[rms]cph()},
\code{\link[rms]Gls()}, \code{\link[rms]summary.rms()}, \code{\link[rms]anova.rms()}.
}

