% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nb_ratioci.R
\name{calc_nb_ratioci}
\alias{calc_nb_ratioci}
\alias{calc_nb_ratioci.default}
\alias{calc_nb_ratioci.mira}
\alias{calc_nb_ratioci.negbin}
\title{Calculate incidence rate ratio and confidence limits from a negative binomial regression
model}
\usage{
calc_nb_ratioci(nbObj = NULL, ...)

\method{calc_nb_ratioci}{default}(nbObj = NULL, nbCoefs, nbVcov, predVar,
  adjustTo = NULL, df, alpha = 0.05)

\method{calc_nb_ratioci}{mira}(nbObj, predVar, adjustTo = NULL, df,
  alpha = 0.05)

\method{calc_nb_ratioci}{negbin}(nbObj, predVar, adjustTo = NULL, df,
  alpha = 0.05)
}
\arguments{
\item{nbObj}{Model fit of class glm.nb, or mice::mira object fit using glm.nb.}

\item{nbCoefs}{Vector of coefficients from a glm.nb.}

\item{nbVcov}{Variance-covariance matrix from a glm.nb fit.}

\item{predVar}{Character string; name of main predictor variable.}

\item{adjustTo}{Numeric vector of length 2; values to adjust predVar to if predVar is continuous.
Defaults to c(25th, 75th percentiles).}

\item{df}{Data frame used to determine type of predVar, get defaults for adjustTo and to
calculate nonlinear terms if applicable.}

\item{alpha}{Alpha level for confidence limits. Defaults to 0.05.}
}
\value{
matrix with one row per comparison, containing columns for point estimate, confidence
limits, and reference and comparison levels used.
}
\description{
The most common way to express incidence rate ratios is simply exp(beta), representing a one-unit
increase in the covariate value; however, in the case of continuous covariates, this is often not
a practically meaningful difference (one year of age among adults, eg, or a one-unit change in
mean arterial pressure). This function allows you to specify a clinically meaningful comparison
in the case of continuous covariates, defaulting to the 75th vs the 25th percentiles of values in
the data frame specified.
}
\details{
Currently does not handle interaction terms.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method used when passing coefficients and variance-covariance matrix,
vs original model fit

Default method uses given coefficients and variance-covariance matrix. Intended use cases include
bootstrapped negative binomial models.

\item \code{mira}: Method for glm.nb models fit with mice objects.

\item \code{negbin}: Method for glm.nb models (no imputation).
}}
\examples{

## Create data frame
df <- data.frame(y = round(rexp(n = 100, rate = 0.5)),
                 v1 = sample(1:5, size = 100, replace = TRUE),
                 v2 = rnorm(n = 100))

## Fit negative binomial model
mymod <- MASS::glm.nb(y ~ v1 * v2, data = df)

calc_nb_ratioci(mymod, predVar = 'v1', df = df)


## Create data frame
df <- data.frame(y = round(rexp(n = 100, rate = 0.5)),
                 v1 = sample(1:5, size = 100, replace = TRUE),
                 v2 = rnorm(n = 100))

## Fit negative binomial model
mymod <- MASS::glm.nb(y ~ v1 * v2, data = df)
mycoefs <- coef(mymod)
myvcov <- vcov(mymod)

calc_nb_ratioci(nbCoefs = mycoefs, nbVcov = myvcov, predVar = 'v1', df = df)


## Create data frame
df <- data.frame(y = round(rexp(n = 100, rate = 0.5)),
                 v1 = sample(c(1:5, NA), size = 100, replace = TRUE),
                 v2 = rnorm(n = 100))

## Impute missing data using mice
mids.df <- mice(df)

## Fit negative binomial model
mymod <- with(mids.df, MASS::glm.nb(y ~ v1 * v2))

calc_nb_ratioci(mymod, predVar = 'v1', df = df)


## Create data frame
df <- data.frame(y = round(rexp(n = 100, rate = 0.5)),
                 v1 = sample(1:5, size = 100, replace = TRUE),
                 v2 = rnorm(n = 100))

## Fit negative binomial model
mymod <- MASS::glm.nb(y ~ v1 * v2, data = df)

calc_nb_ratioci(mymod, predVar = 'v1', df = df)

}
\seealso{
\code{\link[MASS]{glm.nb}}; \code{\link[mice]{mice}}, \code{\link[mice]{pool}};
\code{\link[Hmisc]{rcspline.eval}} for nonlinear terms.

\code{\link[MASS]{glm.nb}}; \code{\link[Hmisc]{rcspline.eval}} for nonlinear terms.

\code{\link[MASS]{glm.nb}}; \code{\link[mice]{mice}} \code{\link[Hmisc]{rcspline.eval}}
for nonlinear terms.

\code{\link[MASS]{glm.nb}}; \code{\link[Hmisc]{rcspline.eval}} for nonlinear terms.
}

