% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms_po_assume.R
\name{rms_po_assume}
\alias{rms_po_assume}
\title{Create figures to visually examine proportional odds assumption from an `lrm` model fit.}
\usage{
rms_po_assume(lrmObj, cuts, plotVars = NULL, mfrowAuto = FALSE, modelData)
}
\arguments{
\item{cuts}{Numeric vector; sequence of points to cut outcome. Should not include lowest outcome
level.}

\item{plotVars}{Character vector; which variables to plot. Defaults to all.}

\item{mfrowAuto}{Logical; whether to determine par(mfrow = ...) automatically. Defaults to FALSE.}

\item{modelData}{Data.frame; data set used to fit original model. Used to fit logistic models
with outcome dichotomized at each cut point.}

\item{rmsObj}{Model fit of class `lrm`.}
}
\description{
Based on code and strategy outlined in Frank Harrell's Regression Modeling Strategies (see
citation below).
}
\examples{

df <- data.frame(ptclass = sample(1:4, size = 20, replace = TRUE),
                 v1 = rnorm(n = 20),
                 v2 = rnorm(mean = 5, sd = 1, n = 20))

mymod <- lrm(ptclass ~ v1 + v2, data = df)
rms_po_assume(mymod, cuts = 2:4, modelData = df)

}
\seealso{
Harrell FE. *Regression Modeling Strategies: with applications to linear models,
logistic regression, and survival analysis.* New York: Springer Science + Business Media, LLC,
2001.
}

